Move privileged and kubelet_etc_hosts into common directory.
